+++ R/aemet_forecast_daily.R
@@ -1,8 +1,18 @@
 aemet_forecast_daily <- function(x, verbose = FALSE) {
   single <- lapply(x, function(x) {
-    res <- aemet_forecast_daily_single(x, verbose = verbose)
+    res <- try(aemet_forecast_daily_single(x, verbose = verbose), silent = TRUE)
+    if (inherits(res, "try-error")) {
+      message(
+        "\nAEMET API call for '", x, "' returned an error\n",
+        "Return NULL for this query"
+      )
+      return(NULL)
+    }
+    return(res)
   })
   bind <- dplyr::bind_rows(single)
+
+  bind <- aemet_hlp_guess(bind, preserve = "id")
 
   return(bind)
 }
@@ -14,9 +24,15 @@ aemet_forecast_daily_single <- function(x, verbose = FALSE) {
       verbose = verbose
     )
 
+  pred$elaborado <- as.POSIXlt(gsub("T", " ", pred$elaborado),
+    tz = "Europe/Madrid"
+  )
+
   # Unnesting this dataset is complex
-  col_types <- vapply(pred, class, FUN.VALUE = character(1))
+  col_types <- get_col_first_class(pred)
   vars <- names(col_types[col_types %in% c("list", "data.frame")])
+
+
 
   first_lev <- tidyr::unnest(pred, col = dplyr::all_of(vars), names_sep = "_")
 
@@ -31,4 +47,15 @@ aemet_forecast_daily_single <- function(x, verbose = FALSE) {
 
   master_end <- dplyr::bind_cols(master, pred_dia)
   return(master_end)
+}
+
+
+# Helper to return first class of column
+
+get_col_first_class <- function(df) {
+  res <- vapply(df, function(x) {
+    return(class(x)[1])
+  }, FUN.VALUE = character(1))
+
+  return(res)
 }