+++ R/aemet_forecast_hourly.R
@@ -1,9 +1,18 @@
 aemet_forecast_hourly <- function(x, verbose = FALSE) {
   single <- lapply(x, function(x) {
-    res <- aemet_forecast_hourly_single(x, verbose = verbose)
+    res <- try(aemet_forecast_hourly_single(x, verbose = verbose), silent = TRUE)
+    if (inherits(res, "try-error")) {
+      message(
+        "\nAEMET API call for '", x, "' returned an error\n",
+        "Return NULL for this query"
+      )
+      return(NULL)
+    }
+    return(res)
   })
   bind <- dplyr::bind_rows(single)
 
+  bind <- aemet_hlp_guess(bind, preserve = "id")
   return(bind)
 }
 
@@ -14,8 +23,12 @@ aemet_forecast_hourly_single <- function(x, verbose = FALSE) {
       verbose = verbose
     )
 
+  pred$elaborado <- as.POSIXlt(gsub("T", " ", pred$elaborado),
+    tz = "Europe/Madrid"
+  )
+
   # Unnesting this dataset is complex
-  col_types <- vapply(pred, class, FUN.VALUE = character(1))
+  col_types <- get_col_first_class(pred)
   vars <- names(col_types[col_types %in% c("list", "data.frame")])
 
   first_lev <- tidyr::unnest(pred, col = dplyr::all_of(vars), names_sep = "_")